requirements.txt
bta_lib==1.0.0
pandas==1.0.3
python_binance==0.7.5
